Update QuickSort.java
I tried to pick a median pivot value, but I ended up just setting the medPivot equal to array[n].  When I took the median of three random pivot values, I got a completely different sort everytime.  But when I followed the pseduocode of the textbook, I was able to get it to sort properly, while probably at the expense of more comparisons.